Make the Closing event public.

+++ test/Renci.SshNet.Tests/Classes/Sftp/Responses/SftpAttrsResponseTest.cs
@@ -28,7 +28,7 @@ namespace Renci.SshNet.Tests.Classes.Sftp.Responses
 
             Assert.IsNull(target.Attributes);
             Assert.AreEqual(_protocolVersion, target.ProtocolVersion);
-            Assert.AreEqual((uint) 0, target.ResponseId);
+            Assert.AreEqual(0U, target.ResponseId);
             Assert.AreEqual(SftpMessageTypes.Attrs, target.SftpMessageType);
         }
